+++ NEWS.md
@@ -4,6 +4,7 @@
 * New argument nullval in make_forest_plot, which adds a vertical reference line at this value. (By default a line is still added at 1 if using log scale.)
 * New argument minse in make_shape_plot and make_forest_plot which sets the minimum standard error to use when scaling point size. This allows scaling to be made consistent between plots.
 * Argument col.keep available in make_forest_plot() to keep columns in the returned data frame.
+* Using the addtext argument of make_forest_plot(), you now need to include an equals or less than sign with the p-value.
 
 # ckbplotr 0.3.0